Mem: Add and implement FlushAll command in Ruby.
TODO: CHANGE USES OF FLUSHALL TO APPROPRIATE ACQUIRE/RELEASE FLAGS/TYPES

A FlushAll message requests the receiving controller to flush everything from
its cache.

This patch:
 - Adds a FlushAllReq/Resp MemCmd.
 - FLUSHALL RubyRequestType
 - Flash invalidate function to the Ruby cache controllers.

The flash invalidate function invalidates all lines which are resident in the
cache. There is a check to be sure that there are no lines in the busy or r/w
state, although this may need to change in the future to support a more
general flash invalidate.
* * *
ruby: Fix flash invalidate memory leak

The flashInvalidate function used to set entry AccessPermission to NotPresent
and delete all tag entries. This resulted in the allocate function finding the
entry for previously flashInvalidated entries, but then just setting the
m_cache pointer to a newly allocated entry and leaking the old entry. This is
a buggy allocate function, which shouldn't overwrite m_cache entry pointers,
but we can avoid this by simply deleting all entries in the flashInvalidate
function.