Mem: Require flush requests to have a response
* * *
Ruby: Fix flush response request leak

The flush-response patch makes the FlushReq MemCmd require a response, which
causes requests to not be deleted in the packet destructor during cooldown of
the caches before checkpointing. Requests are 72B, so larger systems with
reasonably large caches leak a lot of memory for each checkpoint taken.

NOTE: This change will be folded into the jason/flush-responses patch.