ruby: Generalize the Cluster network
This patch adds two generalizations to the Cluster network:
1) In the case that components from one Cluster should be connected to
components within another Cluster, we can add those components to both
Clusters. This creates a problem if these two partially-connected Clusters are
included as sub-Clusters of another Cluster, because the recursive definition
causes the shared portions of the sub-Clusters to be traversed for each path
entering the shared portions. Add a check to see if the sub-Cluster has already
been instantiated, and if so, simply return to the super-Cluster that is
calling the makeTopology function.

2) In the case that multiple sub-Clusters should be connected but the router
within one should not be connected to the top-level Cluster (e.g. a tree
hierarchy with shared components between subtrees), the Cluster would, by
default, add a link between the sub-Cluster router and the top-level Cluster.
Add a variable to specify whether the sub-Cluster's router should be connected
to the top-level Cluster's router. Default the choice to True.