Ruby: Enable GPU-like atomic data updates
To support GPU-like atomic data updates, Ruby needs to differentiate GPU atomic
memory accesses in the RubyPort and Sequencer, and the RubyPort needs to call
out to separate functionality in gem5-gpu that is capable of completing the
memory accesses and associated atomic operations.
* * *
Updated to how atomics are handled.
will be folded into atomic patch